+++ files/aws-tools/snapshot_attach.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys, os, logging, argparse, jmespath, boto3, requests, json, logging, socket, sh, textwrap
+from botocore.config import Config
 from retrying import retry
 from itertools import imap, repeat
 from collections import namedtuple
@@ -411,6 +412,12 @@ if __name__ == '__main__':
     args = parse_args()
     set_logging_level(args.quiet, args.verbose)
     log.debug('Args: %r', args)
+
+    boto3_config = Config(
+        retries = {
+            'max_attempts': 360
+            }
+        )
 
     ec2         = boto3.resource('ec2')
     instance    = ec2_instance()